requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=2.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
camelot-py>=0.11.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.32.0
pytest>=8.0.0

# Auto-chunking script dependencies
pandas>=2.0.0
openpyxl>=3.1.0
